+++ ChangeLog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.9.1] - 2022-09-04
 ### Added
 - `e3nn.reduced_symmetric_tensor_product_basis(irreps: Irreps, order: int)`
 - `e3nn.IrrepsArray.filtered(keep: List[Irrep])`